Removed unused dependencies, adjusted how UI navigation moves, grouped struct fields
