+++ crates/simula_hexgrid/src/hexgrid.rs
@@ -1,25 +1,11 @@
-use crate::{pathfinding::{HexOrientation::*, *}, user_interface::*};
+use crate::{pathfinding::{HexOrientation::*, *}, user_interface::*, worldmap_material::*};
 use bevy::{
-    //core_pipeline::Transparent3d,
-    ecs::system::{lifetimeless::*, SystemParamItem},
     math::prelude::*,
-    pbr::{MeshPipeline, MeshPipelineKey, MeshUniform, SetMeshBindGroup, SetMeshViewBindGroup},
     prelude::*,
     render::{
-        mesh::{GpuBufferInfo, MeshVertexBufferLayout},
-        render_asset::RenderAssets,
-        //render_component::{ExtractComponent, ExtractComponentPlugin},
-        render_phase::{
-            AddRenderCommand, DrawFunctions, EntityRenderCommand, RenderCommandResult, RenderPhase,
-            SetItemPipeline, TrackedRenderPass,
-        },
-        render_resource::*,
-        renderer::RenderDevice,
-        view::{ComputedVisibility, ExtractedView, Msaa, NoFrustumCulling, Visibility},
-        RenderApp, RenderStage, extract_component::{ExtractComponent, ExtractComponentPlugin},
-    }, core_pipeline::core_3d::Transparent3d,
+        view::{ComputedVisibility, NoFrustumCulling, Visibility}
+    }, 
 };
-use bytemuck::{Pod, Zeroable};
 use simula_camera::orbitcam::*;
 use simula_core::prng::*;
 use std::hash::Hash;
@@ -36,8 +22,6 @@ pub struct TempHexTiles;
 
 #[derive(Component)]
 pub struct HexagonTiles;
-
-pub struct DespawnTileEvent;
 
 pub struct RenderPathEvent {
     pub value: RenderAction,
@@ -59,6 +43,11 @@ pub struct NodeStartEnd {
     pub queue_end: (i32, i32),
     pub destination_reached: bool,
     pub queue_node: Vec<((i32, i32), f32, Vec<(i32, i32)>, f32)>,
+    pub nodes: HashMap<(i32, i32), f32>,
+    pub counter_one: i32,
+    pub counter_two: i32,
+    pub shortest_highlight: Vec<(i32, i32)>,
+    pub random_complexity: f32,
     pub nodes_weighted: HashMap<(i32, i32), (f32, f32)>,
     pub node_astar_scores: HashMap<(i32, i32), f32>,
     pub start_weight: f32,
@@ -75,6 +64,11 @@ impl Default for NodeStartEnd {
             queue_end: (2, 3),
             destination_reached: true,
             queue_node: vec![((0, 0), 0.0, Vec::<(i32, i32)>::new(), 0.0)],
+            nodes: HashMap::new(),
+            counter_one: 0,
+            counter_two: 0,
+            shortest_highlight: vec![(1, 2), (2, 3)],
+            random_complexity: 0.0,
             nodes_weighted: HashMap::new(),
             node_astar_scores: HashMap::new(),
             start_weight: 0.0,
@@ -89,17 +83,12 @@ pub struct ShortestPathBuilder {
     pub render_min_row: i32,
     pub render_max_row: i32,
     pub render_size: i32,
-    pub nodes: HashMap<(i32, i32), f32>,
     pub min_column: i32,
     pub max_column: i32,
     pub min_row: i32,
     pub max_row: i32,
     pub tile_coord_x: i32,
     pub tile_coord_z: i32,
-    pub counter_one: i32,
-    pub counter_two: i32,
-    pub shortest_highlight: Vec<(i32, i32)>,
-    pub random_complexity: f32,
 }
 
 impl Default for ShortestPathBuilder {
@@ -110,17 +99,12 @@ impl Default for ShortestPathBuilder {
             render_min_row: -51,
             render_max_row: 77,
             render_size: 128,
-            nodes: HashMap::new(),
             min_column: -1,
             max_column: 2048,
             min_row: -1,
             max_row: 2048,
             tile_coord_x: 0,
             tile_coord_z: 0,
-            counter_one: 0,
-            counter_two: 0,
-            shortest_highlight: vec![(1, 2), (2, 3)],
-            random_complexity: 0.0,
         }
     }
 }
@@ -131,33 +115,35 @@ pub fn hexgrid_viewer(
     mut shortest_path: ResMut<ShortestPathBuilder>,
     mut render_path_event: EventReader<RenderPathEvent>,
     mut orbit_camera: Query<&mut OrbitCamera>,
-    mut despawn_tile_event: EventWriter<DespawnTileEvent>,
+    mut despawn_tile_objects: Query<Entity, (With<HexagonTiles>, Without<TempHexTiles>)>,
+    mut hex_tile_objects: Query<Entity, With<TempHexTiles>>,
+    mut tile_visibility: Query<&mut Visibility, With<TempHexTiles>>,
 ) {
     for mut orbit_camera in orbit_camera.iter_mut() {
         for event in render_path_event.iter() {
             match &event.value {
                 RenderAction::RenderUp => {
                     if shortest_path.render_max_row < 2048 {
-                        shortest_path.render_min_row += shortest_path.render_size;
-                        shortest_path.render_max_row += shortest_path.render_size;
+                        shortest_path.render_min_row += shortest_path.render_size / 2;
+                        shortest_path.render_max_row += shortest_path.render_size / 2;
                     }
                 }
                 RenderAction::RenderDown => {
                     if shortest_path.render_min_row >= 0 {
-                        shortest_path.render_min_row -= shortest_path.render_size;
-                        shortest_path.render_max_row -= shortest_path.render_size;
+                        shortest_path.render_min_row -= shortest_path.render_size / 2;
+                        shortest_path.render_max_row -= shortest_path.render_size / 2;
                     }
                 }
                 RenderAction::RenderLeft => {
                     if shortest_path.render_min_column >= 0 {
-                        shortest_path.render_min_column -= shortest_path.render_size;
-                        shortest_path.render_max_column -= shortest_path.render_size;
+                        shortest_path.render_min_column -= shortest_path.render_size / 2;
+                        shortest_path.render_max_column -= shortest_path.render_size / 2;
                     }
                 }
                 RenderAction::RenderRight => {
                     if shortest_path.render_max_column < 2048 {
-                        shortest_path.render_min_column += shortest_path.render_size;
-                        shortest_path.render_max_column += shortest_path.render_size;
+                        shortest_path.render_min_column += shortest_path.render_size / 2;
+                        shortest_path.render_max_column += shortest_path.render_size / 2;
                     }
                 }
 
@@ -203,7 +189,7 @@ pub fn hexgrid_viewer(
                 .insert(HexgridObject)
                 .insert(TempHexTiles);
 
-            despawn_tile_event.send(DespawnTileEvent);
+                despawn_tile(&mut commands, &mut despawn_tile_objects, &mut hex_tile_objects, &mut tile_visibility);
         }
     }
 }
@@ -214,7 +200,9 @@ pub fn hexgrid_rebuilder(
     mut shortest_path: ResMut<ShortestPathBuilder>,
     mouse_button_input: Res<Input<MouseButton>>,
     orbit_camera: Query<&mut OrbitCamera>,
-    mut despawn_tile_event: EventWriter<DespawnTileEvent>,
+    mut despawn_tile_objects: Query<Entity, (With<HexagonTiles>, Without<TempHexTiles>)>,
+    mut hex_tile_objects: Query<Entity, With<TempHexTiles>>,
+    mut tile_visibility: Query<&mut Visibility, With<TempHexTiles>>,
 ) {
     for orbit_camera in orbit_camera.iter() {
         if mouse_button_input.pressed(MouseButton::Right) {
@@ -260,20 +248,15 @@ pub fn hexgrid_rebuilder(
                 ))
                 .insert(TempHexTiles)
                 .insert(HexgridObject);
-
-            despawn_tile_event.send(DespawnTileEvent);
         }
+        despawn_tile(&mut commands, &mut despawn_tile_objects, &mut hex_tile_objects, &mut tile_visibility)
     }
 }
 
 pub fn select_tile(
     mut maps: Query<&mut HexgridData>,
     shortest_path: ResMut<ShortestPathBuilder>,
-    mut despawn_tile_event: EventReader<DespawnTileEvent>,
-    mut tile_visibility: Query<&mut Visibility, With<TempHexTiles>>,
-    mut commands: Commands,
-    despawn_tile_objects: Query<Entity, (With<HexagonTiles>, Without<TempHexTiles>)>,
-    hex_tile_objects: Query<Entity, With<TempHexTiles>>,
+    node_start_end: ResMut<NodeStartEnd>,
 ) {
     for mut map in maps.iter_mut() {
         if map.len() == shortest_path.render_size as usize * shortest_path.render_size as usize {
@@ -315,241 +298,38 @@ pub fn select_tile(
                         .map(move |z| (x, z))
                 })
                 .enumerate()
-                .filter(|&(_i, x)| shortest_path.shortest_highlight.contains(&x))
+                .filter(|&(_i, x)| node_start_end.shortest_highlight.contains(&x))
                 .for_each(|(i, _x)| {
                     map.0[i].color = Color::hsla(360.0, 1.0, 0.5, 0.1).as_rgba_u32()
                 });
-
-            for _event in despawn_tile_event.iter() {
-                for mut visibility in tile_visibility.iter_mut() {
-                    for ent_despawn in despawn_tile_objects.iter() {
-                        for ent in hex_tile_objects.iter() {
-                            visibility.is_visible = true;
-                            commands
-                                .entity(ent)
-                                .insert(HexagonTiles)
-                                .remove::<TempHexTiles>();
-                            commands.entity(ent_despawn).despawn_recursive();
-                        }
-                    }
-                }
-            }
         }
     }
 }
 
-#[derive(Component, Deref)]
-pub struct HexgridData(pub Vec<HexData>);
-impl ExtractComponent for HexgridData {
-    type Query = &'static HexgridData;
-    type Filter = ();
-
-    fn extract_component(item: bevy::ecs::query::QueryItem<Self::Query>) -> Self {
-        HexgridData(item.0.clone())
-    }
-}
-
-pub struct HexgridMaterialPlugin;
-
-impl Plugin for HexgridMaterialPlugin {
-    fn build(&self, app: &mut App) {
-        app.add_plugin(ExtractComponentPlugin::<HexgridData>::default());
-        app.sub_app_mut(RenderApp)
-            .add_render_command::<Transparent3d, DrawCustom>()
-            .init_resource::<HexgridPipeline>()
-            .init_resource::<SpecializedMeshPipelines<HexgridPipeline>>()
-            .add_system_to_stage(RenderStage::Queue, queue_custom)
-            .add_system_to_stage(RenderStage::Prepare, prepare_instance_buffers);
-    }
-}
-
-#[derive(Clone, Copy, Pod, Zeroable)]
-#[repr(C)]
-pub struct HexData {
-    pub position: Vec3,
-    pub scale: f32,
-    pub color: u32,
-}
-
-#[allow(clippy::too_many_arguments)]
-fn queue_custom(
-    transparent_3d_draw_functions: Res<DrawFunctions<Transparent3d>>,
-    custom_pipeline: Res<HexgridPipeline>,
-    msaa: Res<Msaa>,
-    mut pipelines: ResMut<SpecializedMeshPipelines<HexgridPipeline>>,
-    mut pipeline_cache: ResMut<PipelineCache>,
-    meshes: Res<RenderAssets<Mesh>>,
-    material_meshes: Query<
-        (Entity, &MeshUniform, &Handle<Mesh>),
-        (With<Handle<Mesh>>, With<HexgridData>),
-    >,
-    mut views: Query<(&ExtractedView, &mut RenderPhase<Transparent3d>)>,
+pub fn despawn_tile(
+    commands: &mut Commands,
+    despawn_tile_objects: &mut Query<Entity, (With<HexagonTiles>, Without<TempHexTiles>)>,
+    hex_tile_objects: &mut Query<Entity, With<TempHexTiles>>,
+    tile_visibility: &mut Query<&mut Visibility, With<TempHexTiles>>,
 ) {
-    let draw_custom = transparent_3d_draw_functions
-        .read()
-        .get_id::<DrawCustom>()
-        .unwrap();
-
-    let msaa_key = MeshPipelineKey::from_msaa_samples(msaa.samples);
-
-    for (view, mut transparent_phase) in views.iter_mut() {
-        let view_matrix = view.transform.compute_matrix();
-        let view_row_2 = view_matrix.row(2);
-        for (entity, mesh_uniform, mesh_handle) in material_meshes.iter() {
-            if let Some(mesh) = meshes.get(mesh_handle) {
-                let key =
-                    msaa_key | MeshPipelineKey::from_primitive_topology(mesh.primitive_topology);
-                let pipeline = pipelines
-                    .specialize(&mut pipeline_cache, &custom_pipeline, key, &mesh.layout)
-                    .unwrap();
-                transparent_phase.add(Transparent3d {
-                    entity,
-                    pipeline,
-                    draw_function: draw_custom,
-                    distance: view_row_2.dot(mesh_uniform.transform.col(3)),
-                });
-            }
+    for mut visibility in tile_visibility.iter_mut() {
+        for ent_despawn in despawn_tile_objects.iter_mut() {
+            commands.entity(ent_despawn).despawn_recursive();
         }
-    }
-}
-
-#[derive(Component)]
-pub struct InstanceBuffer {
-    buffer: Buffer,
-    length: usize,
-}
-
-fn prepare_instance_buffers(
-    mut commands: Commands,
-    query: Query<(Entity, &HexgridData)>,
-    render_device: Res<RenderDevice>,
-) {
-    for (entity, instance_data) in query.iter() {
-        let buffer = render_device.create_buffer_with_data(&BufferInitDescriptor {
-            label: Some("instance data buffer"),
-            contents: bytemuck::cast_slice(instance_data.as_slice()),
-            usage: BufferUsages::VERTEX | BufferUsages::COPY_DST,
-        });
-        commands.entity(entity).insert(InstanceBuffer {
-            buffer,
-            length: instance_data.len(),
-        });
-    }
-}
-
-pub struct HexgridPipeline {
-    shader: Handle<Shader>,
-    mesh_pipeline: MeshPipeline,
-}
-
-impl FromWorld for HexgridPipeline {
-    fn from_world(world: &mut World) -> Self {
-        let world = world.cell();
-        let asset_server = world.get_resource::<AssetServer>().unwrap();
-        asset_server.watch_for_changes().unwrap();
-        let shader = asset_server.load("shaders/hexgrid.wgsl");
-
-        let mesh_pipeline = world.get_resource::<MeshPipeline>().unwrap();
-
-        HexgridPipeline {
-            shader,
-            mesh_pipeline: mesh_pipeline.clone(),
+        for ent in hex_tile_objects.iter() {
+            visibility.is_visible = true;
+            commands
+                .entity(ent)
+                .insert(HexagonTiles)
+                .remove::<TempHexTiles>();
         }
-    }
-}
-
-impl SpecializedMeshPipeline for HexgridPipeline {
-    type Key = MeshPipelineKey;
-
-    fn specialize(
-        &self,
-        key: Self::Key,
-        layout: &MeshVertexBufferLayout,
-    ) -> Result<RenderPipelineDescriptor, SpecializedMeshPipelineError> {
-        let mut descriptor = self.mesh_pipeline.specialize(key, layout)?;
-        descriptor.vertex.shader = self.shader.clone();
-        descriptor.vertex.buffers.push(VertexBufferLayout {
-            array_stride: std::mem::size_of::<HexData>() as u64,
-            step_mode: VertexStepMode::Instance,
-            attributes: vec![
-                VertexAttribute {
-                    format: VertexFormat::Float32x4,
-                    offset: 0,
-                    shader_location: 3, // shader locations 0-2 are taken up by Position, Normal and UV attributes
-                },
-                VertexAttribute {
-                    format: VertexFormat::Uint32,
-                    offset: VertexFormat::Float32x4.size(),
-                    shader_location: 4,
-                },
-            ],
-        });
-        descriptor.fragment.as_mut().unwrap().shader = self.shader.clone();
-        descriptor.layout = Some(vec![
-            self.mesh_pipeline.view_layout.clone(),
-            self.mesh_pipeline.mesh_layout.clone(),
-        ]);
-
-        Ok(descriptor)
-    }
-}
-
-type DrawCustom = (
-    SetItemPipeline,
-    SetMeshViewBindGroup<0>,
-    SetMeshBindGroup<1>,
-    DrawMeshInstanced,
-);
-
-pub struct DrawMeshInstanced;
-impl EntityRenderCommand for DrawMeshInstanced {
-    type Param = (
-        SRes<RenderAssets<Mesh>>,
-        SQuery<Read<Handle<Mesh>>>,
-        SQuery<Read<InstanceBuffer>>,
-    );
-    #[inline]
-    fn render<'w>(
-        _view: Entity,
-        item: Entity,
-        (meshes, mesh_query, instance_buffer_query): SystemParamItem<'w, '_, Self::Param>,
-        pass: &mut TrackedRenderPass<'w>,
-    ) -> RenderCommandResult {
-        let mesh_handle = mesh_query.get(item).unwrap();
-        let instance_buffer = instance_buffer_query.get_inner(item).unwrap();
-
-        let gpu_mesh = match meshes.into_inner().get(mesh_handle) {
-            Some(gpu_mesh) => gpu_mesh,
-            None => return RenderCommandResult::Failure,
-        };
-
-        pass.set_vertex_buffer(0, gpu_mesh.vertex_buffer.slice(..));
-        pass.set_vertex_buffer(1, instance_buffer.buffer.slice(..));
-
-        match &gpu_mesh.buffer_info {
-            GpuBufferInfo::Indexed {
-                buffer,
-                index_format,
-                count,
-            } => {
-                pass.set_index_buffer(buffer.slice(..), 0, *index_format);
-                pass.draw_indexed(0..*count, 0, 0..instance_buffer.length as u32);
-            }
-            GpuBufferInfo::NonIndexed { vertex_count } => {
-                pass.draw(0..*vertex_count, 0..instance_buffer.length as u32);
-            }
-        }
-        RenderCommandResult::Success
     }
 }
 
 pub fn hexagon_pathfinder(
     node_start_end: &mut ResMut<NodeStartEnd>,
     shortest_path: &mut ResMut<ShortestPathBuilder>,
-    //mut calculate_path_event: EventReader<CalculatePathEvent>,
-    //orientation: HexOrientation
 ) {
-    //for _event in calculate_path_event.iter() {
         // you are here
         let start_node = (node_start_end.startx, node_start_end.starty);
 
@@ -562,7 +342,7 @@ pub fn hexagon_pathfinder(
         if node_start_end.destination_reached == true {
             node_start_end.nodes_weighted = HashMap::new();
             // calculate a weighting for each node based on its distance from the end node
-            for (k, v) in shortest_path.nodes.iter() {
+            for (k, v) in node_start_end.nodes.clone().iter() {
                 node_start_end.nodes_weighted.insert(
                     k.to_owned(),
                     (
@@ -699,8 +479,7 @@ pub fn hexagon_pathfinder(
         best_path.push(end_node);
         let best = best_path;
 
-        shortest_path.shortest_highlight = best.clone();
-    //}
+        node_start_end.shortest_highlight = best.clone();
 }
 
 pub struct HexgridPlugin;
@@ -709,14 +488,13 @@ impl Plugin for HexgridPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugin(OrbitCameraPlugin)
             .add_plugin(HexgridMaterialPlugin)
-            .add_event::<DespawnTileEvent>()
             .add_event::<RenderPathEvent>()
             .insert_resource(NodeStartEnd::default())
             .insert_resource(ShortestPathBuilder::default())
             .add_system(ui_system_pathfinding_window)
             .add_system(ui_render_next_tiles)
-            .add_system(select_tile)
-            .add_system(hexgrid_viewer)
-            .add_system(hexgrid_rebuilder);
+            .add_system(select_tile.after(hexgrid_viewer))
+            .add_system(hexgrid_viewer.after(ui_render_next_tiles))
+            .add_system(hexgrid_rebuilder.after(select_tile));
     }
 }
