+++ crates/simula_hexgrid/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod hexgrid;
 pub mod pathfinding;
 pub mod user_interface;
+pub mod worldmap_material;
